agentes/requirements.txt
python-dotenv
grpcio
grpcio-tools
pyyaml
jsonschema
requests
redis
psycopg2-binary
pyjwt
prometheus-client
